pyproject.toml
[project]
name = "unipack-server"
version = "1.0.0"
description = "UniPack API - Odoo Integration Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "email-validator>=2.0.0",
    "google-generativeai>=0.3.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.26.0",
]

[project.scripts]
unipack = "uvicorn:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.26.0",
]
